sc_spatial

sc_spatial

sc_spatial(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  density = FALSE,
  grid.n = 100,
  joint = FALSE,
  joint.fun = prod,
  common.legend = TRUE,
  point.size = 5,
  ...
)

# S4 method for Seurat
sc_spatial(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  density = FALSE,
  grid.n = 100,
  joint = FALSE,
  joint.fun = prod,
  common.legend = TRUE,
  point.size = 5,
  ...
)

# S4 method for SingleCellExperiment
sc_spatial(
  object,
  features = NULL,
  sample.id = NULL,
  image.id = NULL,
  slot = "data",
  image.plot = TRUE,
  image.first.operation = "rotate",
  image.rotate.degree = NULL,
  image.mirror.axis = NULL,
  remove.point = FALSE,
  mapping = NULL,
  ncol = 6,
  density = FALSE,
  grid.n = 100,
  joint = FALSE,
  joint.fun = prod,
  common.legend = TRUE,
  point.size = 5,
  ...
)

Arguments

object

Seurat object

features

selected features to be visualized

sample.id

the index name of sample id, which only work with SingleCellExperiment or SpatialExperiment.

image.id

the index name of image id, which only work with SingleCellExperiment or SpatialExperiment.

slot

if plotting a feature, which data will be used (e.g., 'data', 'counts'), the assay name if object is SingleCellExperiment or SpatialExperiment.

image.plot

whether to display the issue image as background.

image.first.operation

character which the first operation to image, 'rotate' or 'mirror', default is 'rotate'.

image.rotate.degree

integer the degree to ratate image, default is NULL.

image.mirror.axis

character the direction to mirror the image, default is 'h'.

remove.point

whether to remove the spot points, it is nice if your just view the issue image, default is FALSE.

mapping

aesthetic mapping, default is NULL.

ncol

integer number of facet columns if 'length(features) > 1', default is 6.

density

whether plot the 2D weighted kernel density, default is FALSE.

grid.n

number of grid points in the two directions to estimate 2D weighted kernel density, default is 100.

joint

whether joint the multiple features with joint.fun, default is FALSE.

joint.fun

how to joint the multiple features if joint = TRUE, default is prod.

common.legend

whether to use facet_wrap to display the multiple features, default is TRUE.

point.size

the size of point, default is 5.

...

additional parameters.

Value

ggplot object

Examples

if (FALSE) {
library(STexampleData)
# create ExperimentHub instance
eh <- ExperimentHub()
# query STexampleData datasets
myfiles <- query(eh, "STexampleData")
spe <- myfiles[["EH7538"]]
spe <- spe[, colData(spe)$in_tissue == 1]
set.seed(123)
genes <- rownames(spe) |> sample(6) 
p <- sc_spatial(spe, features = genes, 
                image.rotate.degree = -90, 
                image.mirror.axis = NULL, 
                ncol = 3)
}